requirements.txt
cohere>=5.11.0
qdrant-client>=1.11.0
python-docx>=1.1.0
PyPDF2>=3.0.0
numpy>=1.26.0
streamlit>=1.39.0
python-dotenv>=1.0.0
reportlab>=4.0.0
jupyter>=1.0.0
